Resolve multiple on toast subscriptions
The cleanup function didn't seem to be cleaning up correctly (or at
least as I expected), causing each save to increase the amount of toasts
shown. Just threw the listening to happen in that file instead of
wrapped in an `onToast` function used in a `useEffect` (which was a bit
pointless anyway, but I thought might fit some pattern in the app).