Reuse maps
The component wasn't cleaning up properly and causing stacking
`useEffect`s. I don't think this was manifesting in any BIG problems,
but I think was the source of the canvas (I think) errors that were
popping up in the console. Regardless it's a nicer UX to not load map
tiles every time the tab loads.