More tweaks (#10)
* Resolve multiple on toast subscriptions

The cleanup function didn't seem to be cleaning up correctly (or at
least as I expected), causing each save to increase the amount of toasts
shown. Just threw the listening to happen in that file instead of
wrapped in an `onToast` function used in a `useEffect` (which was a bit
pointless anyway, but I thought might fit some pattern in the app).

* Coerce to strings

When the data was loaded, if a field (ex. serial number) was entered as
only a number then it would fail validation.

* Trigger form after resetting data

* Reuse maps

The component wasn't cleaning up properly and causing stacking
`useEffect`s. I don't think this was manifesting in any BIG problems,
but I think was the source of the canvas (I think) errors that were
popping up in the console. Regardless it's a nicer UX to not load map
tiles every time the tab loads.

* Create menu through a side effect import

* Add save menu item

* Disable edit menu when form is disabled

This might help with indicating that the edit actions are just for
editing text in the inputs rather than the images (I'm mainly concerned
about someone thinking undoing after save is possible).

* Fix tests

* Lint