Fixing broken things (#17)
* Get in a submittable state without needing to disable form first

Something made it so until the form was disabled at least once it
wouldn't let you save changes.
Not really sure what the issue was, but fuck it. Logging
`form.formState.isValid` also resolved this for some reason.

* Parse dates as just strings

One of my photos was returning `                   ` (so blank for each
part of the date format), and I couldn't come up with a quick and simple
way of handling potential blank but not `null`/`undefined` values
without blowing up.

* Forward logging to terminal for debugging release

* Remove white space for fields

One of my pictures had white space for a date field that wasn't touched,
so failed to save since that's not valid for exiftool.

* Lint

* Switch mode to onChange